Clean up code and return UUID from rollbar response
This simplifies the code by combining the shared logic across functions
into a new `reportErrorSWithOptions` that optionally returns a rollbar
uuid to help better identify errors. There is further cleanup that could
be done but this seemed like a decent place to stop. This is a mildly
breaking change (I aligned the arguments to the functions so they match)
but largely tries to keep compatibility.